# services/CoordinatorService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Job Aggregator Coordinator Service.",
        "title": "Vaca Coordinator API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Yehor Reil",
            "url": "https://github.com/ReilEgor"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
        "/sources": {
            "get": {
                "description": "Retrieve a list of all job boards and platforms supported by the aggregator",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Get all available job sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_rest_handlers.ListSourcesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error with domain.ErrFailedToGetSources",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tasks": {
            "post": {
                "description": "Initialize a job search process across multiple sources using specific keywords",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create a new scraping task",
                "parameters": [
                    {
                        "description": "Task Configuration",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_rest_handlers.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_rest_handlers.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Retrieve the current progress and status of a specific task by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get status of a scraping task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_rest_handlers.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Task ID is required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vacancies": {
            "get": {
                "description": "Get a paginated list of job vacancies filtered by title, source, or keywords",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vacancies"
                ],
                "summary": "Search and filter vacancies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by job title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by job source (e.g., djinni, dou)",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_rest_handlers.SearchVacanciesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_transport_rest_handlers.CreateTaskRequest": {
            "type": "object",
            "required": [
                "keywords",
                "sources"
            ],
            "properties": {
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_transport_rest_handlers.CreateTaskResponse": {
            "type": "object",
            "required": [
                "created_at",
                "status",
                "task_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
        "internal_transport_rest_handlers.ListSourcesResponse": {
            "type": "object",
            "required": [
                "sources",
                "total"
            ],
            "properties": {
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transport_rest_handlers.SourceResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "internal_transport_rest_handlers.SearchVacanciesResponse": {
            "type": "object",
            "required": [
                "items",
                "total"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transport_rest_handlers.VacancyResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "internal_transport_rest_handlers.SourceResponse": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_transport_rest_handlers.VacancyResponse": {
            "type": "object",
            "required": [
                "about",
                "company",
                "id",
                "link",
                "requirements",
                "title"
            ],
            "properties": {
                "about": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "requirements": {
                    "type": "string"
                },
                "salary": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "minLength": 3
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}