# services/CoordinatorService/internal/transport/rest/handlers/source_handler.go
package handler

import (
	"log/slog"
	"net/http"

	"github.com/ReilEgor/Vaca/services/CoordinatorService/internal/domain"
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type SourceResponse struct {
	ID   uuid.UUID `json:"id" binding:"required"`
	Name string    `json:"name" binding:"required"`
}

type ListSourcesResponse struct {
	Sources []SourceResponse `json:"sources" binding:"required,dive"`
	Total   int64            `json:"total" binding:"required,gte=0"`
}

// GetAvailableSources godoc
// @Summary      Get all available job sources
// @Description  Retrieve a list of all job boards and platforms supported by the aggregator
// @Tags         sources
// @Produce      json
// @Success      200  {object}  ListSourcesResponse
// @Failure      500  {object}  map[string]string "Internal server error with domain.ErrFailedToGetSources"
// @Router       /sources [get]
func (h *Handler) GetAvailableSources(c *gin.Context) {
	ctx := c.Request.Context()

	sources, total, err := h.uc.GetAvailableSources(ctx)
	if err != nil {
		h.logger.Error("failed to get available sources", slog.Any("error", err))
		c.JSON(http.StatusInternalServerError, gin.H{"error": domain.ErrFailedToGetSources.Error()})
		return
	}

	respSources := make([]SourceResponse, len(sources))
	for i, s := range sources {
		respSources[i] = SourceResponse{
			ID:   s.ID,
			Name: s.Name,
		}
	}

	c.JSON(http.StatusOK, ListSourcesResponse{
		Sources: respSources,
		Total:   total,
	})
}